+++ sample/src/App.tsx
@@ -60,15 +60,17 @@ Sentry.init({
 const Stack = createStackNavigator();
 
 const App = () => {
-  const navigation = React.createRef<NavigationContainerRef>();
-
-  React.useEffect(() => {
-    reactNavigationV5Instrumentation.registerNavigationContainer(navigation);
-  }, []);
+  const navigation = React.useRef<NavigationContainerRef>();
 
   return (
     <Provider store={store}>
-      <NavigationContainer ref={navigation}>
+      <NavigationContainer
+        ref={navigation}
+        onReady={() => {
+          reactNavigationV5Instrumentation.registerNavigationContainer(
+            navigation,
+          );
+        }}>
         <Stack.Navigator>
           <Stack.Screen name="Home" component={HomeScreen} />
           <Stack.Screen name="Tracker" component={TrackerScreen} />